requirements.txt
asgiref~=3.2.10
asn1crypto==0.24.0
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
cffi==1.14.3
channels~=2.4.0
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.0
Django~=3.1
django-extensions==3.0.9
geographiclib==1.50
geopy==2.0.0
hyperlink==20.0.1
idna==2.6
importlib-metadata==2.0.0
incremental==17.5.0
keyring==10.6.0
keyrings.alt==3.0
Markdown~=3.2.2
Pillow==8.0.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pygobject==3.26.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-magic==0.4.18
pytz==2020.4
pyxdg==0.25
redis==3.5.3
SecretStorage==2.3.1
service-identity==18.1.0
six==1.11.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
uWSGI==2.0.19.1
zipp==3.4.0
zope.interface==5.1.2

asgiref~=3.2.10
django-parler~=2.2
django-model-utils~=4.1.1